requirements.txt
-e .
pytest
aws-cdk.core
aws-cdk.aws_ec2
